requirements.txt
websocket-client
evdev
